+++ sdk/core/Azure.Core/src/RequestContext.cs
@@ -1,0 +1,39 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Threading;
+using Azure.Core;
+using Azure.Core.Pipeline;
+
+namespace Azure
+{
+    /// <summary>
+    /// Options which can be used to control the behavior of a request sent by a client.
+    /// </summary>
+    public class RequestContext
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="RequestContext"/> class.
+        /// </summary>
+        public RequestContext()
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="RequestContext"/> class using the given <see cref="ErrorOptions"/>.
+        /// </summary>
+        /// <param name="options"></param>
+        public static implicit operator RequestContext(ErrorOptions options) => new RequestContext { ErrorOptions = options };
+
+        /// <summary>
+        /// The token to check for cancellation.
+        /// </summary>
+        public CancellationToken CancellationToken { get; set; } = CancellationToken.None;
+
+        /// <summary>
+        /// Controls under what conditions the operation raises an exception if the underlying response indicates a failure.
+        /// </summary>
+        public ErrorOptions ErrorOptions { get; set; } = ErrorOptions.Default;
+    }
+}