+++ sdk/network/Azure.ResourceManager.Network/tests/Tests/FlowLogTests.cs
@@ -44,18 +44,18 @@ namespace Azure.ResourceManager.Network.Tests
             var networkSecurityGroup = new NetworkSecurityGroupData() { Location = location, };
 
             // Put Nsg
-            var securityGroupContainer = resourceGroup.GetNetworkSecurityGroups();
-            var putNsgResponseOperation = await securityGroupContainer.CreateOrUpdateAsync(networkSecurityGroupName, networkSecurityGroup);
+            var securityGroupCollection = resourceGroup.GetNetworkSecurityGroups();
+            var putNsgResponseOperation = await securityGroupCollection.CreateOrUpdateAsync(networkSecurityGroupName, networkSecurityGroup);
             await putNsgResponseOperation.WaitForCompletionAsync();;
             // Get NSG
-            Response<NetworkSecurityGroup> getNsgResponse = await securityGroupContainer.GetAsync(networkSecurityGroupName);
+            Response<NetworkSecurityGroup> getNsgResponse = await securityGroupCollection.GetAsync(networkSecurityGroupName);
 
             string networkWatcherName = Recording.GenerateAssetName("azsmnet");
             var properties = new NetworkWatcherData { Location = location };
 
             //Create network Watcher
-            var networkWatcherContainer = resourceGroup.GetNetworkWatchers();
-            await networkWatcherContainer.CreateOrUpdateAsync(networkWatcherName, properties);
+            var networkWatcherCollection = resourceGroup.GetNetworkWatchers();
+            await networkWatcherCollection.CreateOrUpdateAsync(networkWatcherName, properties);
 
             //Create storage
             string storageName = Recording.GenerateAssetName("azsmnet");
@@ -95,10 +95,10 @@ namespace Azure.ResourceManager.Network.Tests
             //};
 
             //configure flowlog and TA
-            //var configureFlowLog1Operation = await networkWatcherContainer.Get(networkWatcherName).Value.SetFlowLogConfigurationAsync(configParameters);
+            //var configureFlowLog1Operation = await networkWatcherCollection.Get(networkWatcherName).Value.SetFlowLogConfigurationAsync(configParameters);
             //await configureFlowLog1Operation.WaitForCompletionAsync();;
 
-            //var queryFlowLogStatus1Operation = await networkWatcherContainer.Get(networkWatcherName).Value.GetFlowLogStatusAsync(new FlowLogStatusParameters(getNsgResponse.Value.Id));
+            //var queryFlowLogStatus1Operation = await networkWatcherCollection.Get(networkWatcherName).Value.GetFlowLogStatusAsync(new FlowLogStatusParameters(getNsgResponse.Value.Id));
             //Response<FlowLogInformation> queryFlowLogStatus1 = await queryFlowLogStatus1Operation.WaitForCompletionAsync();;
             ////check both flowlog and TA config and enabled status
             //Assert.AreEqual(queryFlowLogStatus1.Value.TargetResourceId, configParameters.TargetResourceId);
@@ -116,10 +116,10 @@ namespace Azure.ResourceManager.Network.Tests
 
             ////disable TA
             //configParameters.FlowAnalyticsConfiguration.NetworkWatcherFlowAnalyticsConfiguration.Enabled = false;
-            //var configureFlowLog2Operation = await networkWatcherContainer.Get(networkWatcherName).Value.SetFlowLogConfigurationAsync(configParameters);
+            //var configureFlowLog2Operation = await networkWatcherCollection.Get(networkWatcherName).Value.SetFlowLogConfigurationAsync(configParameters);
             //await configureFlowLog2Operation.WaitForCompletionAsync();;
 
-            //var queryFlowLogStatus2Operation = await networkWatcherContainer.Get(networkWatcherName).Value.GetFlowLogStatusAsync(new FlowLogStatusParameters(getNsgResponse.Value.Id));
+            //var queryFlowLogStatus2Operation = await networkWatcherCollection.Get(networkWatcherName).Value.GetFlowLogStatusAsync(new FlowLogStatusParameters(getNsgResponse.Value.Id));
             //Response<FlowLogInformation> queryFlowLogStatus2 = await queryFlowLogStatus2Operation.WaitForCompletionAsync();;
 
             ////check TA disabled and ensure flowlog config is unchanged
@@ -130,10 +130,10 @@ namespace Azure.ResourceManager.Network.Tests
 
             ////disable flowlog (and TA)
             //configParameters.Enabled = false;
-            //var configureFlowLog3Operation = await networkWatcherContainer.Get(networkWatcherName).Value.SetFlowLogConfigurationAsync(configParameters);
+            //var configureFlowLog3Operation = await networkWatcherCollection.Get(networkWatcherName).Value.SetFlowLogConfigurationAsync(configParameters);
             //await configureFlowLog3Operation.WaitForCompletionAsync();;
 
-            //var queryFlowLogStatus3Operation = await networkWatcherContainer.Get(networkWatcherName).Value.GetFlowLogStatusAsync(new FlowLogStatusParameters(getNsgResponse.Value.Id));
+            //var queryFlowLogStatus3Operation = await networkWatcherCollection.Get(networkWatcherName).Value.GetFlowLogStatusAsync(new FlowLogStatusParameters(getNsgResponse.Value.Id));
             //Response<FlowLogInformation> queryFlowLogStatus3 = await queryFlowLogStatus3Operation.WaitForCompletionAsync();;
 
             ////check both flowlog and TA disabled