+++ sdk/resources/Azure.ResourceManager.Resources/tests/Scenario/ApplicationOperationsTests.cs
@@ -19,15 +19,16 @@ namespace Azure.ResourceManager.Resources.Tests
         [RecordedTest]
         public async Task Delete()
         {
+            Subscription subscription = await Client.GetDefaultSubscriptionAsync();
             string rgName = Recording.GenerateAssetName("testRg-5-");
             ResourceGroupData rgData = new ResourceGroupData(Location.WestUS2);
-            var lro = await Client.DefaultSubscription.GetResourceGroups().CreateOrUpdateAsync(rgName, rgData);
+            var lro = await subscription.GetResourceGroups().CreateOrUpdateAsync(rgName, rgData);
             ResourceGroup rg = lro.Value;
             string appDefName = Recording.GenerateAssetName("appDef-D-");
             ApplicationDefinitionData appDefData = CreateApplicationDefinitionData(appDefName);
             ApplicationDefinition appDef = (await rg.GetApplicationDefinitions().CreateOrUpdateAsync(appDefName, appDefData)).Value;
             string appName = Recording.GenerateAssetName("application-D-");
-            ApplicationData applicationData = CreateApplicationData(appDef.Id, Client.DefaultSubscription.Id + Recording.GenerateAssetName("/resourceGroups/managed-5-"), Recording.GenerateAssetName("s5"));
+            ApplicationData applicationData = CreateApplicationData(appDef.Id, subscription.Id + Recording.GenerateAssetName("/resourceGroups/managed-5-"), Recording.GenerateAssetName("s5"));
             Application application = (await rg.GetApplications().CreateOrUpdateAsync(appName, applicationData)).Value;
             await application.DeleteAsync();
             var ex = Assert.ThrowsAsync<RequestFailedException>(async () => await application.GetAsync());