+++ sdk/webpubsub/Microsoft.Azure.WebJobs.Extensions.WebPubSub/CHANGELOG.md
@@ -1,6 +1,26 @@
 # Release History
 
-## 1.0.0 (Unreleased)
+## 1.1.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 1.0.0 (2021-11-09)
+
+### Breaking Changes
+- Move output binding object to namespace `Microsoft.Azure.WebJobs.Extensions.WebPubSub`.
+- Add a few static create methods under `WebPubSubAction` to easily discover available actions.
+- Add dependency `Azure.Messageing.WebPubSub` for service calls.
+- Rename output binding from `WebPubSubOperation` to `WebPubSubAction`, and add `Action` as suffix for derived classes.
+- Rename `operationKind` to `actionName` in output binding for javascript to deserialize.
+- Changed the name and type of the `Uri` properties in `WebPubSubConnection` to match other libraries in the Azure SDK for .NET.
+
+## 1.0.0-beta.4 (2021-11-09)
 
 ### Features Added
 - Add `Connection` attribute to input binding and trigger binding to support upstream validations.
@@ -14,8 +34,6 @@
 
 ### Bugs Fixed
 - Fix json deserialize issues and limited to string only to reduce ambiguity.
-
-### Other Changes
 
 ## 1.0.0-beta.3 (2021-07-26)